Do not generate changelog for release
Previously, the release workflow generated a changelog from the commit history and added it to the release notes.

The generated changelog serves no value. If a user wants to see a list of commits, a higher quality list can be obtained
by looking at the commit history.

Furthermore, a generated changelog is actually very harmful in that it encourages the project maintainer to shirk their
duty (which is specified in the formal release procedure) of writing a real changelog.

Since the changelog generation provides no value, the generation code is hereby removed from the release workflow.