Configure grep commands in installation script to use POSIX extended syntax
A script is provided to facilitate installation of Arduino Lint.

The grep utility is used for several purposes in the script.

grep supports multiple flavors of regular expression syntax:

* "basic" (BRE)
* "extended" (ERE)
* "perl" (PCRE)

The default "BRE" syntax is inconsistent with the syntax used in all other common regular expression applications:

> In basic regular expressions the meta-characters ?, +, {, |, (, and ) lose their special meaning; instead use the
> backslashed versions \?, \+, \{, \|, \(, and \).

The "PCRE" syntax that would otherwise be preferred is not available when using grep on BSD/macOS machines.

So grep should be configured to use the "ERE" in scripts that are intended to be cross-platform (as is the case with
this script).