Improve version number regex in installation script
The installation script scrapes the GitHub releases page for the version number of the latest release. A regular
expression is used to extract the version number from the title.

Previously, the regular expression would return a corrupted version number if the name of the repository owner or
repository contained numbers. For example, if used in a repository owned by GitHub user `per1234`, where the latest
release version was 1.2.3, it would return the following version number:

```
1.2.3
1234
```